Moved the add_action/add_filter to above the callback.
Why? It improves readability.

Neither PHP nor WordPress care about the order.  It runs whether
the hook is above or below its callback.

But it does matter for readability, which is vital component of
code quality.

Why does it matter?

Humans read down the page, not up and backwards.  As we open a
file, we begin to read it.  With the add_action/add_filter after
the function, we first read the function itself.

Is it a callback? We don't know until we read down the page and find
where it's hooked to an event.  The _intent_ of this function is not
being clearly communicated.

By flipping it around, you can immediately recognize that this
function is a callback that's hooked into this event name.

Readability empowers anyone to pop into the project, quickly
figure out what's going on, and then get to work.

Our job as developers is to communicate our intent by making code
"human readable."

Learn more @see https://knowthecode.io/add-action-before-function