Returned early if the post type is not post.
The first conditional is a "no go" condition, where we
do not want to process our business logic if it is not a "post"
post type.  We do not want "no go" conditions to continue to flow
as that is a flawed design.  Rather, when we find one, we want to
immediately stop and bail out.

The other benefits are:

1. It communicates the intent that if this is not a "post" post type,
then bail out.
2. Only do the work (business logic) when it's a "post" post type.
3. It unwraps the business logic, moving it to the left. That makes it
more readable as there are less levels (depths) of indentation.
4. It prevents someone from misunderstanding what this "no go" is.
Imagine that someone comes along and puts code after the if code block.
That would cause unexpected behavior.

@see https://torquemag.io/2018/04/code-review-part-1-fixing-design-flaw-return-early-strategy/