Modernized the plugin to use PHP Namespace.
Why?

1. As of PHP version 5.3.0, we have namespaces.

2. Namespaces behave like JavaScript namespaces, allowing us to protect
the global space.

3. It eliminates the prefixing to improve readability.
Prefixing makes a function harder to read and quickly understand.

Why?

We have to read past the namespace to understand what this function
will do when we call it.  The prefix only exists to provide a unique
name and avoid naming conflicts.  It does not serve the developer.

4. It packages up our code by our Brand\PackageName to give the
file and code more contextual meaning.  We can quickly identify
who built the package and which package it is.

5. Namespaces protect us better from naming conflicts.  The changes
of another plugin, theme, or package having the same combination of
namespace words is much lower than a prefix.