Refactored redundant file loaders.
In the plugin's bootstrap file, there are 4 files to be loaded.
Previously the code repeated this pattern:

include plugin_dir_path( __FILE__ ) . 'lib/filename.php';

What's wrong with that pattern?

Notice that the `plugin_dir_path( __FILE__ )` function is being
called 4 times.  When we go to look at the code for that function,
you'll trace it to find that it also calls:

- dirname()
- trailingslashit()
- untrailingslashit()
- rtrim()

Each time we need to load a file, there are 5 functions that will
run.  Add that up:

4 files to load x 5 functions = 20 functions total.

That's both redundant and less performant.

-----------------------
How do we improve this?
-----------------------

Option 1:

We could store the result of the plugin's dir path into a variable,
e.g. ``$plugin_dir = plugin_dir_path( __FILE__ );`.

That's a bad strategy.  Why?  It's more work for PHP.  That makes
it slower and less performant.

1. PHP has to assign the plugin's root directory to a variable.
That assignment requires it to do this:

- add the variable into the symbols table
- bind the variable to the directory value's memory location

2. Every time we use the variable, PHP has to go to its symbols table
and look up the value.

Option 2:

We can use use __DIR__, which is equivalent to dirname( __FILE__ );

Why is this better?  The file has to exist in the root of
the plugin folder, meaning it's already in the root.  There's no
need to look up the dirname() and then process the trailing slash.

Instead, we can use PHP's magic constant __DIR__.

@see http://php.net/manual/en/language.constants.predefined.php#constant.dir