Refactored path to plugin's dir and URI.
The assets need the plugin's root directory as well as it's URI.
The previous design was slower.  Why?  It had to call `plugins_url()`
and `plugin_dir_path()` 4 times each.  These functions call other
functions.

As we previously discussed, the function `plugin_dir_path()` calls 4
other functions.  Let's do the math:

4 times it runs x 5 functions = 20 total functions being called.

How about `plugins_url()`?  This function is expensive as it calls:

- wp_normalize_path() 3 times
- strpos()
- set_url_scheme()
- plugin_basename()
- dirname()
- is_string() 2 times
- ltrim() 2 times
- apply_filters()
- and makes 8 assignment operations

That's slow.

We can improve the performance by using 2 functions in the bootstrap
file of the plugin:

- _get_plugin_url()
- _get_plugin_directory()

These functions are internal to the plugin and, therefore, start with
an underscore to communicate that intent.

This strategy does add 2 more functions and as well as a static
variable and memory location. But it is much faster than using the
old approach, as discussed above.