Fixed potential error when no recent posts.
The function `wp_get_recent_posts()` returns either an array or
false.  The value of `false` is not countable.

Also, the function `count()` is slower than checking if the value is
empty.  Here `empty` will return `true` when the value is false, 0,
[], null, 0.0, '', or '0';

This is a faster approach and more expressive to the intent.